Repositories: move implementations to HMS/Reposistories/Doctorine
