+++ app/Providers/RepositoryServiceProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Providers;
+namespace Providers;
 
 use HMS\Entities\Meta;
 use HMS\Entities\Role;
@@ -11,7 +11,10 @@ use HMS\Repositories\RoleRepository;
 use HMS\Repositories\UserRepository;
 use HMS\Repositories\InviteRepository;
 use Illuminate\Support\ServiceProvider;
-use HMS\Repositories\DoctrineUserRepository;
+use HMS\Repositories\Doctrine\DoctrineMetaRepository;
+use HMS\Repositories\Doctrine\DoctrineRoleRepository;
+use HMS\Repositories\Doctrine\DoctrineUserRepository;
+use HMS\Repositories\Doctrine\DoctrineInviteRepository;
 
 class RepositoryServiceProvider extends ServiceProvider
 {
@@ -33,15 +36,15 @@ class RepositoryServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->singleton(MetaRepository::class, function ($app) {
-            return new MetaRepository($app['em'], $app['em']->getClassMetaData(Meta::class));
+            return new DoctrineMetaRepository($app['em'], $app['em']->getClassMetaData(Meta::class));
         });
 
         $this->app->singleton(InviteRepository::class, function ($app) {
-            return new InviteRepository($app['em'], $app['em']->getClassMetaData(Invite::class));
+            return new DoctrineInviteRepository($app['em'], $app['em']->getClassMetaData(Invite::class));
         });
 
         $this->app->singleton(RoleRepository::class, function ($app) {
-            return new RoleRepository($app['em'], $app['em']->getClassMetaData(Role::class));
+            return new DoctrineRoleRepository($app['em'], $app['em']->getClassMetaData(Role::class));
         });
 
         $this->app->singleton(UserRepository::class, function ($app) {
